ServerWarden/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
python-dateutil==2.8.2
pytz==2024.1 